Keep arguments containing spaces in one piece
Replacing $@ with "$@" doesn't break arguments into pieces if/where
they contain spaces. We also have to quote arguments when they are
first scanned into the shell variable ARGS. There may be a more
elegant solution: just use a `for` and the "$@" in the first place to
avoid breaking them apart at all.